requirements.txt
alembic==1.7.6
click==8.0.4
dataclasses==0.6
Flask==2.0.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
fuzzywuzzy==0.18.0
gevent==1.4.0
greenlet==0.4.16
gunicorn==20.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
psycogreen==1.0.2
psycopg2-binary==2.9.3
python-Levenshtein==0.12.2
SQLAlchemy==1.4.31
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
requests==2.27.1
pytest==7.0.1
pytest-html==3.1.1